+++ src/Server/authentication.cpp
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   authentication.cpp                                 :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: amalangi <amalangi@student.42.fr>          +#+  +:+       +#+        */
+/*   By: rwintgen <rwintgen@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/05 22:17:18 by amalangi          #+#    #+#             */
-/*   Updated: 2024/10/09 21:01:35 by amalangi         ###   ########.fr       */
+/*   Updated: 2024/10/24 11:37:29 by rwintgen         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,7 +21,7 @@ void Server::authentication(int fd, const char *buffer)
 	Client &client = getClientByFd(fd);
 	if (!client.isAuth())
 	{
-		for (int i = 0; i < parser.message.size(); i++)
+		for (size_t i = 0; i < parser.message.size(); i++)
 		{
 			if ((parser.message[i][0] == "NICK" || parser.message[i][0] == "USER") && parser.message[i].size() > 1) {
 				if (parser.message[i][1].find('#') != std::string::npos) {
@@ -30,7 +30,7 @@ void Server::authentication(int fd, const char *buffer)
 					disconnectClientByFd(fd);
 					return ;
 				}
-				for (int j = 0; j < this->_clientList.size(); j++) {
+				for (size_t j = 0; j < this->_clientList.size(); j++) {
 					if ((parser.message[i][1] == "NICK" ? this->_clientList[j]->getNick() : this->_clientList[j]->getUser()) == parser.message[i][1]) {
 						sendMessage(fd, ":MyChell.beer 433 ERR_NICKNAMEINUSE " + parser.message[i][1] + ": Nickname is already in use\r\n");
 						std::cout << "\e[1;31m" << "Client <" << fd << "> Disconnected for ERR_NICKNAMEINUSE !" << "\e[0;37m" << std::endl;
