+++ src/Server/command/names.cpp
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   names.cpp                                          :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: amalangi <amalangi@student.42.fr>          +#+  +:+       +#+        */
+/*   By: rwintgen <rwintgen@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/06 00:43:51 by amalangi          #+#    #+#             */
-/*   Updated: 2024/10/08 02:43:26 by amalangi         ###   ########.fr       */
+/*   Updated: 2024/10/24 11:20:08 by rwintgen         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -26,7 +26,7 @@ void Server::Command_NAMES(int fd, std::vector<std::string> msg,Client &client) 
     }
 	std::vector<Client *>users = channel->getClientList();
 	std::string clientListString;
-	for (int i = 0; i < users.size(); i++) {
+	for (size_t i = 0; i < users.size(); i++) {
 		clientListString += (channel->isOp(*users[i]) ? "@" : "") + users[i]->getNick() + " ";
 	}
 	sendMessage(fd, ":MyChell.beer 353 " + client.getNick() + " = " + channelName + " :" + clientListString + "\r\n");
